requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
httpx
PyMuPDF
openai
python-dotenv
azure-storage-blob
pyodbc
pymssql
azure-storage-blob
psycopg2-binary
pydantic>=2.0
